safari-534.53-branch/LayoutTests/fast/dom/undetectable-style-filter.html

<html>
<head>
<script>
function print(message, color) 
{
    var paragraph = document.createElement("div");
    paragraph.appendChild(document.createTextNode(message));
    paragraph.style.fontFamily = "monospace";
    if (color)
        paragraph.style.color = color;
    document.getElementById("console").appendChild(paragraph);
}

function shouldBe(a, b)
{
    var evalA;
    try {
        evalA = eval(a);
    } catch(e) {
        evalA = e;
    }
    
    if (evalA == b)
        print("PASS: " + a + " should be " + b + " and is.", "green");
    else
        print("FAIL: " + a + " should be " + b + " but instead is " + evalA + ".", "red");
}

function test() 
{
    if (window.layoutTestController)
        layoutTestController.dumpAsText();
    
    shouldBe("typeof document.body.style.filter", "undefined");
    shouldBe("document.body.style.filter ? true : false", false);
    shouldBe("document.body.style.filter == undefined", true);
}
</script>
</head>

<body onload="test();">
<p>This page tests whether the CSS 'filter' property is undetectable. If the test passes, you'll see a series of 'PASS' messages below.</p>
<hr>

<div id='console'></div>

</body>
</html>
